refactored; moved code to session.py and model_client.py
